refactor(web): extract task polling into useTaskPoller module
